+++ example/lib/leyoujia/button.dart
@@ -29,9 +29,9 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
     super.dispose();
   }
 
-  void _showSelectedResult(DropdownSelectorResult result) {
+  void _showSelectedResult(SelectorEntries selected) {
     final l10n = AppLocalizations.of(context);
-    final conditions = '${result.selected.flatten()}';
+    final conditions = '${selected.flatten()}';
     ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(
         content: Text(l10n?.filterUpdated ?? ''),
@@ -63,16 +63,17 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
     );
   }
 
-  HouseFilter? _dropdownSelectorResultParser(DropdownSelectorResult result) {
+  HouseFilter? _dropdownSelectorResultParser(
+      DropdownTabData tabData, SelectorEntries selected) {
     final filter = HouseFilter(cityId: userCityId);
-    if (result.tabIndex == 0) {
+    if (tabData.index == 0) {
       // 区域
-      _filtersRepo.regionResult = result.selected;
-      final category = result.selected.firstOrNull;
+      _filtersRepo.regionResult = selected;
+      final category = selected.firstOrNull;
       if (category == null) return null;
       if (category.id == 'region') {
         // 行政区
-        filter.district = result
+        filter.district = selected
             .cascadingPairsOf('region')
             .map((p) => {
                   "district_id": p.id,
@@ -81,7 +82,7 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
             .toList(growable: false);
       } else if (category.id == 'metro') {
         // 地铁
-        filter.metro = result
+        filter.metro = selected
             .cascadingPairsOf('metro')
             .map((p) => {
                   "line_id": p.id,
@@ -91,18 +92,18 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
       } else if (category.id == 'nearby') {
         // 附近
         final nearbyRadiusMeters =
-            result.findIdsAtLevel(category, 1).firstOrNull;
+            selected.findIdsAtLevel(category, 1).firstOrNull;
         filter.nearbyRadiusMeters = nearbyRadiusMeters;
         filter.userLatLon = userLatLon;
       }
-    } else if (result.tabIndex == 1) {
+    } else if (tabData.index == 1) {
       // 价格筛选
-      _filtersRepo.buyPriceResult = result.selected;
-      final category = result.selected.firstOrNull;
+      _filtersRepo.buyPriceResult = selected;
+      final category = selected.firstOrNull;
       if (category == null) return null;
       if (category.id == 'total') {
         // 总价
-        filter.totalPrice = result
+        filter.totalPrice = selected
             .childRangesOf('total')
             .map((e) => {
                   "id": e.id,
@@ -113,7 +114,7 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
       }
       if (category.id == 'unit') {
         // 单价
-        filter.unitPrice = result
+        filter.unitPrice = selected
             .childRangesOf('unit')
             .map((e) => {
                   "id": e.id,
@@ -122,13 +123,13 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
                 })
             .toList(growable: false);
       }
-    } else if (result.tabIndex == 2) {
+    } else if (tabData.index == 2) {
       // 户型筛选
-      _filtersRepo.floorPlanBuyResult = result.selected;
-      filter.livingRoom = result.childIdsOf('living_room');
-      filter.bathroom = result.childIdsOf('bathroom');
-      filter.balcony = result.childIdsOf('balcony');
-      filter.area = result
+      _filtersRepo.floorPlanBuyResult = selected;
+      filter.livingRoom = selected.childIdsOf('living_room');
+      filter.bathroom = selected.childIdsOf('bathroom');
+      filter.balcony = selected.childIdsOf('balcony');
+      filter.area = selected
           .childRangesOf('area')
           .map((e) => {
                 "id": e.id,
@@ -136,27 +137,28 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
                 "max": e.max,
               })
           .toList(growable: false);
-    } else if (result.tabIndex == 3) {
+    } else if (tabData.index == 3) {
       // 更多筛选
-      _filtersRepo.moreBuyResult = result.selected;
-      filter.homeType = result.childIdsOf('home_type');
-      filter.saleStatus = result.childIdsOf('sale_status');
-      filter.openTime = result.childIdsOf('open_time');
-      filter.deliveryTime = result.childIdsOf('delivery_time');
-      filter.decorationStatus = result.childIdsOf('decoration_status');
-      filter.buildingFeatures = result.childIdsOf('building_features');
-      filter.houseViewService = result.childIdsOf('house_view_service');
-    } else if (result.tabIndex == 4) {
+      _filtersRepo.moreBuyResult = selected;
+      filter.homeType = selected.childIdsOf('home_type');
+      filter.saleStatus = selected.childIdsOf('sale_status');
+      filter.openTime = selected.childIdsOf('open_time');
+      filter.deliveryTime = selected.childIdsOf('delivery_time');
+      filter.decorationStatus = selected.childIdsOf('decoration_status');
+      filter.buildingFeatures = selected.childIdsOf('building_features');
+      filter.houseViewService = selected.childIdsOf('house_view_service');
+    } else if (tabData.index == 4) {
       // 排序筛选
-      _filtersRepo.sortBuyResult = result.selected;
-      filter.sort = result.firstSelectedId;
+      _filtersRepo.sortBuyResult = selected;
+      filter.sort = selected.firstSelectedId;
     }
     return filter;
   }
 
-  void _handleSelectorChange(DropdownSelectorResult result) async {
+  void _handleSelectorChange(
+      DropdownTabData tabData, SelectorEntries selected) async {
     final l10n = AppLocalizations.of(context);
-    _filter = _dropdownSelectorResultParser(result);
+    _filter = _dropdownSelectorResultParser(tabData, selected);
     if (_filter == null) {
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(content: Text(l10n?.filterParseFailed ?? '')),
@@ -165,9 +167,9 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
     }
   }
 
-  void _handleSelectorApply(DropdownSelectorResult result) {
+  void _handleSelectorApply(DropdownTabData tabData, SelectorEntries selected) {
     final l10n = AppLocalizations.of(context);
-    _filter = _dropdownSelectorResultParser(result);
+    _filter = _dropdownSelectorResultParser(tabData, selected);
     if (_filter == null) {
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(content: Text(l10n?.filterParseFailed ?? '')),
@@ -198,15 +200,15 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
                 return MyCheckbox(value: selected);
               },
             ),
-            onChanged: (result) {
-              debugPrint('onChanged: $result');
-              _handleSelectorChange(result);
-              _showSelectedResult(result);
+            onChanged: (tabData, selected) {
+              debugPrint('onChanged: $tabData $selected');
+              _handleSelectorChange(tabData, selected);
+              _showSelectedResult(selected);
             },
-            onApplied: (result) {
-              debugPrint('onApplied: $result');
-              _handleSelectorApply(result);
-              _showSelectedResult(result);
+            onApplied: (tabData, selected) {
+              debugPrint('onApplied: $selected');
+              _handleSelectorApply(tabData, selected);
+              _showSelectedResult(selected);
             },
             onReset: () {
               debugPrint('onReset');
@@ -239,15 +241,15 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
                 ),
                 applyText: AppLocalizations.of(context)?.apply ?? '',
               ),
-              onChanged: (result) {
-                debugPrint('onChanged: $result');
-                _handleSelectorChange(result);
-                _showSelectedResult(result);
+              onChanged: (tabData, selected) {
+                debugPrint('onChanged: $selected');
+                _handleSelectorChange(tabData, selected);
+                _showSelectedResult(selected);
               },
-              onApplied: (result) {
-                debugPrint('onApplied: $result');
-                _handleSelectorApply(result);
-                _showSelectedResult(result);
+              onApplied: (tabData, selected) {
+                debugPrint('onApplied: $selected');
+                _handleSelectorApply(tabData, selected);
+                _showSelectedResult(selected);
               },
               onReset: () {
                 debugPrint('onReset');
@@ -270,15 +272,15 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
                 mainAxisSpacing: 8,
                 sideBarTheme: const SelectorSideBarTheme(width: 98),
               ),
-              onChanged: (result) {
-                debugPrint('onChanged: $result');
-                _handleSelectorChange(result);
-                _showSelectedResult(result);
+              onChanged: (tabData, selected) {
+                debugPrint('onChanged: $selected');
+                _handleSelectorChange(tabData, selected);
+                _showSelectedResult(selected);
               },
-              onApplied: (result) {
-                debugPrint('onApplied: $result');
-                _handleSelectorApply(result);
-                _showSelectedResult(result);
+              onApplied: (tabData, selected) {
+                debugPrint('onApplied: $selected');
+                _handleSelectorApply(tabData, selected);
+                _showSelectedResult(selected);
               },
               onReset: () {
                 debugPrint('onReset');
@@ -299,15 +301,15 @@ class _ButtonDemoPageState extends State<ButtonDemoPage> {
                   return MyRadio(value: selected);
                 },
               ),
-              onChanged: (result) {
-                debugPrint('onChanged: $result');
-                _handleSelectorChange(result);
-                _showSelectedResult(result);
+              onChanged: (tabData, selected) {
+                debugPrint('onChanged: $selected');
+                _handleSelectorChange(tabData, selected);
+                _showSelectedResult(selected);
               },
-              onApplied: (result) {
-                debugPrint('onApplied: $result');
-                _handleSelectorApply(result);
-                _showSelectedResult(result);
+              onApplied: (tabData, selected) {
+                debugPrint('onApplied: $selected');
+                _handleSelectorApply(tabData, selected);
+                _showSelectedResult(selected);
               },
               onReset: () {
                 debugPrint('onReset');